+++ src/core/Draw.h
@@ -26,7 +26,7 @@ namespace Draw {
 
     struct SpriteData {
         const uint8_t* pixel_data;
-        uint16_t pixel_data_size;
+        uint32_t pixel_data_size;
         int width;
         int height;
         int src_x;
@@ -61,10 +61,10 @@ namespace Draw {
               int scale = 1, int z_index = 1,
               const FontData* font = &Font::DEFAULT_BLANK);
     void rect(int x, int y, int width, int height, uint32_t color, bool fill = true, int thickness = 1, int z_index = 1);
-    void sprite(int x, int y, const uint8_t* pixel_data, uint16_t pixel_data_size, int width, int height, int z_index = 1);
+    void sprite(int x, int y, const uint8_t* pixel_data, uint32_t pixel_data_size, int width, int height, int z_index = 1);
     void sprite_frame(
         int screen_x, int screen_y,
-        const uint8_t* sheet_pixels, uint16_t sheet_pixels_size,
+        const uint8_t* sheet_pixels, uint32_t sheet_pixels_size,
         int sheet_width, int sheet_height,
         int src_x, int src_y, int src_w, int src_h,
         int z_index = 1
